# server/controllers/Mrtproject_db/generated/ScheduleControllerGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE scheduleControllerGenerated.js PLEASE EDIT ../scheduleController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import ScheduleModel from "../../../models/Mrtproject_db/ScheduleModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import ScheduleController from "../ScheduleController";

const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/schedule`;
    router.post(baseUrl + "", authorize([]), ScheduleController.create);
    router.delete(baseUrl + "/:id", authorize([]), ScheduleController.delete);
    router.get(baseUrl + "/findBy_karyawan/:key", authorize([]), ScheduleController.findBy_karyawan);
    router.get(baseUrl + "/findBy_shift/:key", authorize([]), ScheduleController.findBy_shift);
    router.get(baseUrl + "/:id", authorize([]), ScheduleController.get);
    router.get(baseUrl + "", authorize([]), ScheduleController.list);
    router.post(baseUrl + "/:id", authorize([]), ScheduleController.update);
  },


  // CRUD METHODS


  /**
  * scheduleModel.create
  *   @description CRUD ACTION create
  *
  */
  create: async (req, res) => {
    try {
      const result = await ScheduleModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * scheduleModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id
  *
  */
  delete: async (req, res) => {
    try {
      const result = await ScheduleModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * scheduleModel.findBy_karyawan
  *   @description CRUD ACTION findBy_karyawan
  *   @param Objectid key Id della risorsa _karyawan da cercare
  *
  */
  findBy_karyawan: async (req, res) => {
    try {
      const result = await ScheduleModel.findBy_karyawan(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * scheduleModel.findBy_shift
  *   @description CRUD ACTION findBy_shift
  *   @param Objectid key Id della risorsa _shift da cercare
  *
  */
  findBy_shift: async (req, res) => {
    try {
      const result = await ScheduleModel.findBy_shift(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * scheduleModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id resource
  *
  */
  get: async (req, res) => {
    try {
      const result = await ScheduleModel.get(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * scheduleModel.list
  *   @description CRUD ACTION list
  *
  */
  list: async (req, res) => {
    try {
      const result = await ScheduleModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * scheduleModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id
  *
  */
  update: async (req, res) => {
    try {
      const result = await ScheduleModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  // Custom APIs

};

export default generatedControllers;